+++ pkg/exemplars/prometheus.go
@@ -4,12 +4,14 @@
 package exemplars
 
 import (
+	"context"
 	"net/url"
 
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/thanos-io/thanos/pkg/exemplars/exemplarspb"
 	"github.com/thanos-io/thanos/pkg/promclient"
 	"github.com/thanos-io/thanos/pkg/store/labelpb"
+	"github.com/thanos-io/thanos/pkg/tracing"
 )
 
 // Prometheus implements exemplarspb.Exemplars gRPC that allows to fetch exemplars from Prometheus.
@@ -41,7 +43,12 @@ func (p *Prometheus) Exemplars(r *exemplarspb.ExemplarsRequest, s exemplarspb.Ex
 	for _, e := range exemplars {
 		// Make sure the returned series labels are sorted.
 		e.SetSeriesLabels(labelpb.ExtendSortedLabels(e.SeriesLabels.PromLabels(), extLset))
-		if err := s.Send(&exemplarspb.ExemplarsResponse{Result: &exemplarspb.ExemplarsResponse_Data{Data: e}}); err != nil {
+
+		var err error
+		tracing.DoInSpan(s.Context(), "send_exemplars_response", func(_ context.Context) {
+			err = s.Send(&exemplarspb.ExemplarsResponse{Result: &exemplarspb.ExemplarsResponse_Data{Data: e}})
+		})
+		if err != nil {
 			return err
 		}
 	}