+++ CHANGELOG.md
@@ -14,7 +14,8 @@ We use _breaking :warning:_ to mark changes that are not backward compatible (re
 
 ### Added
 
--
+- [#4299](https://github.com/thanos-io/thanos/pull/4299) Tracing: Add tracing to exemplar APIs.
+- [#4327](https://github.com/thanos-io/thanos/pull/4327) Add environment variable substitution to all YAML configuration flags.
 
 ### Fixed
 
@@ -24,7 +25,13 @@ We use _breaking :warning:_ to mark changes that are not backward compatible (re
 
 -
 
-## [v0.21.0](https://github.com/thanos-io/thanos/releases/tag/v0.21.0) - 2021.05.28
+## [v0.21.1](https://github.com/thanos-io/thanos/releases/tag/v0.21.1) - 2021.06.04
+
+### Fixed
+
+- [#4308](https://github.com/thanos-io/thanos/pull/4308) Sidecar: reloader: fix output config file permission
+
+## [v0.21.0](https://github.com/thanos-io/thanos/releases/tag/v0.21.0) - 2021.06.03
 
 ### Added